Populated effects table

Generate populated effects table with calculated CIs

populated_effects_table(data)

Arguments

data

(data.frame) dataset

Value

Effects tble with calculated CIs

Examples

effects_table <- populated_effects_table(effects_table)
#> [2024-07-25 12:14:09] >
#> absolute risk CI for binary outcomes is calculated and saved
#> [2024-07-25 12:14:09] >
#> CI for relative risk for binary outcomes is calculated
#> [2024-07-25 12:14:09] >
#> CI for odds ratio for binary outcomes is calculated and saved
#> [2024-07-25 12:14:09] >
#> CI for treatment difference in continuous outcomes is calculated
#> [2024-07-25 12:14:09] >
#> CI for treatment difference in exposure-adjusted rates
#> is calculated and saved in a dataframe
#> [2024-07-25 12:14:09] >
#> CI for treatment difference in exposure-adjusted rates
#> is calculated and saved in a dataframe
#> [2024-07-25 12:14:09] > Generate populated effects table with calculated CIs
head(effects_table)
#>    Factor     Grouped_Outcome            Outcome                Statistics
#> 1 Benefit Clinical Assessment   Primary Efficacy     % Achieving Remission
#> 2 Benefit Clinical Assessment Secondary Efficacy Mean Change from Baseline
#> 3 Benefit     Quality of Life HR Quality of Life Mean Change from Baseline
#> 4    Risk       Adverse Event     Reoccurring AE                Event Rate
#> 5    Risk       Adverse Event           Rare SAE            Incidence Rate
#> 6    Risk            Toxicity              Liver            Incidence Rate
#>         Type Rate_Type Outcome_Status Filter Category   Trt1 nSub1   N1 Prop1
#> 1     Binary      <NA>     Identified   None      All Drug A   300 1000 0.460
#> 2 Continuous      <NA>     Identified   None      All Drug A    NA 1000    NA
#> 3 Continuous      <NA>     Identified   None      All Drug A    NA 1000    NA
#> 4     Binary EventRate     Identified   None      All Drug A   300 1000 0.190
#> 5     Binary   IncRate     Identified   None      All Drug A    15 1000 0.015
#> 6     Binary   IncRate      Potential   None      All Drug A     4 1000 0.004
#>   Dur1 100PYAR1 IncRate1 nEvent1 100PEY1 EventRate1 Mean1 Se1 Sd1 Drug_Status
#> 1  365       NA       NA      NA      NA         NA    NA  NA  NA    Approved
#> 2  365       NA       NA      NA      NA         NA    65  NA  63    Approved
#> 3   NA       NA       NA      NA      NA         NA    60  NA  60    Approved
#> 4  365       NA       NA     750    1000       0.75    NA  NA  NA    Approved
#> 5  365     1000    0.300      NA      NA         NA    NA  NA  NA    Approved
#> 6  365     1000    0.015      NA      NA         NA    NA  NA  NA    Approved
#>      Trt2 nSub2   N2 Prop2 Dur2 100PYAR2 IncRate2 nEvent2 100PEY2 EventRate2
#> 1 Placebo    50 1000 0.050  365       NA       NA      NA      NA         NA
#> 2 Placebo    NA 1000    NA   NA       NA       NA      NA      NA         NA
#> 3 Placebo    NA 1000    NA   NA       NA       NA      NA      NA         NA
#> 4 Placebo    30 1000 0.030  365       NA       NA      30    1000       0.03
#> 5 Placebo     1 1000 0.002  365     1000    0.001      NA      NA         NA
#> 6 Placebo     1 1000 0.001  365     1000    0.001      NA      NA         NA
#>   Mean2 Se2 Sd2 Diff_LowerCI Diff_UpperCI Diff_IncRate_LowerCI
#> 1    NA  NA  NA  0.376285201  0.443714799                   NA
#> 2    20  NA  16 40.968890317 49.031109683                   NA
#> 3     9  NA   8 47.246045616 54.753954384                   NA
#> 4    NA  NA  NA  0.133486099  0.186513901                   NA
#> 5    NA  NA  NA  0.004973475  0.021026525          0.264995896
#> 6    NA  NA  NA -0.001375156  0.007375156          0.006160144
#>   Diff_IncRate_UpperCI Diff_EventRate_LowerCI Diff_EventRate_UpperCI
#> 1                   NA                     NA                     NA
#> 2                   NA                     NA                     NA
#> 3                   NA                     NA                     NA
#> 4                   NA              0.6652612              0.7747388
#> 5           0.33300410                     NA                     NA
#> 6           0.02183986                     NA                     NA
#>   RelRisk_LowerCI RelRisk_UpperCI OddsRatio_LowerCI OddsRatio_UpperCI
#> 1       6.9644608       12.153130        11.8664287          22.07574
#> 2              NA              NA                NA                NA
#> 3              NA              NA                NA                NA
#> 4       4.3530706        9.214441         5.1033367          11.27156
#> 5       1.7196172       32.710769         1.7332246          33.31626
#> 6       0.4478673       35.724866         0.4476421          35.95848
#>   MCDA_Weight Population Data_Source Quality Notes
#> 1          NA         NA          NA      NA    NA
#> 2          NA         NA          NA      NA    NA
#> 3          NA         NA          NA      NA    NA
#> 4          NA         NA          NA      NA    NA
#> 5          NA         NA          NA      NA    NA
#> 6          NA         NA          NA      NA    NA